move line scanner into separate package
